+++ CHANGELOG.md
@@ -13,6 +13,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Treat scheduled outgoing activities where sender is not specified as invalid.
 - Removed unhashed OAuth tokens from database.
 
+### Removed
+
+- Don't read instance key from `instance_rsa_key` file.
+
 ## [3.23.0] - 2025-04-15
 
 ### Added